+++ packages/types/src/entities/import-job.ts
@@ -1,33 +1,5 @@
-import type {
-  AccountId,
-  AccountPurgeRequestId,
-  BlobId,
-  BoardMessageId,
-  BucketId,
-  ChannelId,
-  CustomFrontId,
-  ExportRequestId,
-  FieldDefinitionId,
-  FieldValueId,
-  FrontingCommentId,
-  FrontingSessionId,
-  GroupId,
-  ImportEntityRefId,
-  ImportJobId,
-  JournalEntryId,
-  MemberId,
-  MessageId,
-  NoteId,
-  PollId,
-  SystemId,
-  TimerId,
-} from "./ids.js";
-import type { UnixMillis } from "./timestamps.js";
-
-// ── Import/Export types ─────────────────────────────────────────────
-// Types for data import, export, and account management.
-// Import payloads use plain string IDs and number timestamps
-// (external format shapes, not internal branded types).
+import type { AccountId, ImportJobId, SystemId } from "../ids.js";
+import type { UnixMillis } from "../timestamps.js";
 
 // ── PluralKit import payloads ───────────────────────────────────────
 
@@ -219,148 +191,6 @@ export interface ImportCheckpointStateV2 {
 }
 
 export type ImportCheckpointState = ImportCheckpointStateV2;
-
-interface ImportEntityRefBase {
-  readonly id: ImportEntityRefId;
-  readonly accountId: AccountId;
-  readonly systemId: SystemId;
-  readonly source: ImportSourceFormat;
-  /** Opaque identifier from the source system (e.g., Mongo ObjectId for SP). */
-  readonly sourceEntityId: string;
-  readonly importedAt: UnixMillis;
-}
-
-/**
- * Maps each ImportEntityType to the branded Pluralscape ID it resolves to.
- * Used to type-scope `ImportEntityRef.pluralscapeEntityId` via a
- * discriminated union. "switch" and "unknown" resolve to raw string because
- * no dedicated branded IDs exist for those categories.
- */
-export interface ImportEntityTargetIdMap {
-  readonly member: MemberId;
-  readonly group: GroupId;
-  readonly "custom-front": CustomFrontId;
-  readonly "fronting-session": FrontingSessionId;
-  readonly "fronting-comment": FrontingCommentId;
-  readonly switch: string;
-  readonly "custom-field": FieldDefinitionId;
-  readonly "field-definition": FieldDefinitionId;
-  readonly "field-value": FieldValueId;
-  readonly note: NoteId;
-  readonly "journal-entry": JournalEntryId;
-  readonly "chat-message": MessageId;
-  readonly "board-message": BoardMessageId;
-  readonly "channel-category": ChannelId;
-  readonly channel: ChannelId;
-  readonly poll: PollId;
-  readonly timer: TimerId;
-  readonly "privacy-bucket": BucketId;
-  readonly "system-profile": SystemId;
-  readonly "system-settings": SystemId;
-  readonly unknown: string;
-}
-
-/**
- * A source-entity to target-entity mapping recorded during an import.
- * Enables idempotent re-imports and cross-device dedup.
- *
- * Discriminated on `sourceEntityType` so consumers get the correct branded
- * target ID via narrowing (no manual cast).
- */
-export type ImportEntityRef = {
-  [K in ImportEntityType]: ImportEntityRefBase & {
-    readonly sourceEntityType: K;
-    readonly pluralscapeEntityId: ImportEntityTargetIdMap[K];
-  };
-}[ImportEntityType];
-
-// ── Export types ─────────────────────────────────────────────────────
-
-/** Status of an export request. */
-export type ExportRequestStatus = "pending" | "processing" | "completed" | "failed";
-
-/** Format for an export manifest. */
-export type ExportFormat = "json" | "csv";
-
-/** Sections that can be included in an export. */
-export type ExportSection =
-  | "members"
-  | "groups"
-  | "fronting-history"
-  | "journal"
-  | "custom-fields"
-  | "notes"
-  | "chat"
-  | "board-messages"
-  | "privacy-buckets"
-  | "system-structure"
-  | "settings"
-  | "polls"
-  | "lifecycle-events";
-
-/** Base type for downloadable report artifacts. */
-export interface DownloadableReport {
-  readonly generatedAt: UnixMillis;
-  readonly sizeBytes: number;
-  readonly downloadUrl: string;
-  readonly expiresAt: UnixMillis;
-}
-
-/** Manifest describing an export package. */
-export interface ExportManifest extends DownloadableReport {
-  readonly systemId: SystemId;
-  readonly format: ExportFormat;
-  readonly sections: readonly ExportSection[];
-}
-
-/** An export request tracking record. */
-export interface ExportRequest {
-  readonly id: ExportRequestId;
-  readonly accountId: AccountId;
-  readonly systemId: SystemId;
-  readonly format: ExportFormat;
-  readonly status: ExportRequestStatus;
-  readonly blobId: BlobId | null;
-  readonly createdAt: UnixMillis;
-  readonly updatedAt: UnixMillis;
-  readonly completedAt: UnixMillis | null;
-}
-
-// ── Account management ──────────────────────────────────────────────
-
-/** Status of an account purge request. */
-export type AccountPurgeStatus = "pending" | "confirmed" | "processing" | "completed" | "cancelled";
-
-/** Format for member reports. */
-export type ReportFormat = "html" | "pdf";
-
-/** Request to purge an entire account and all associated data. */
-export interface AccountPurgeRequest {
-  readonly id: AccountPurgeRequestId;
-  readonly accountId: AccountId;
-  readonly status: AccountPurgeStatus;
-  readonly confirmationPhrase: string;
-  readonly requestedAt: UnixMillis;
-  readonly confirmedAt: UnixMillis | null;
-  readonly scheduledPurgeAt: UnixMillis;
-  readonly completedAt: UnixMillis | null;
-  readonly cancelledAt: UnixMillis | null;
-}
-
-/** A downloadable report of a single member's data. */
-export interface MemberReport extends DownloadableReport {
-  readonly systemId: SystemId;
-  readonly memberId: MemberId;
-  readonly bucketId: BucketId;
-  readonly format: ReportFormat;
-}
-
-/** A "Meet our system" overview report for sharing with friends/family. */
-export interface SystemOverviewReport extends DownloadableReport {
-  readonly systemId: SystemId;
-  readonly bucketId: BucketId;
-  readonly format: ReportFormat;
-}
 
 // ── Canonical SSOT enum tuples ──────────────────────────────────────
 // Imported by @pluralscape/db, @pluralscape/validation, and API router
